refactor: (step4) 당첨 여부를 판별하는 메서드의 위치 변경
- 변경 사유 : Lottos 객체에서 당첨 여부 판별을 위해, 인자로 받은 WinningsLotto 객체의 필드 접근을 위한 get 메서드 제거
- 변경 내용 : Lottos 객체의 match 메서드를 WinningsLotto 객체로 이동
- 변경 효과 :
  - get 메서드 사용 없이 변경 전과 동일한 기능 제공
  - WinningsLotto 객체 필드 접근을 위해 구현한 get 메서드 삭제
- 리팩토링에 따른 코드 변경 :
  - match 메서드에 대한 TC 위치 이동
    - AS-IS : Lottos 객체의 Test Class
    - TO-BE : WinningsLotto 객체의 Test Class
  - 불필요 파일 삭제
