+++ src/main/java/step3/lotto/domain/lotto/Lottos.java
@@ -14,14 +14,6 @@ public class Lottos {
         this.lottos = lottos;
     }
 
-    public MatchStatistic match(Winnings winnings) {
-        MatchStatistic matchStatistic = new MatchStatistic();
-        for (Lotto lotto : lottos) {
-            matchStatistic.add(lotto.match(winnings));
-        }
-        return matchStatistic;
-    }
-
     public List<Lotto> getLottos() {
         return lottos;
     }
