+++ src/main/java/step3/lotto/service/LottoService.java
@@ -1,7 +1,6 @@
 package step3.lotto.service;
 
 import step3.lotto.domain.customer.Customer;
-import step3.lotto.domain.lotto.Lottos;
 import step3.lotto.domain.lotto.MatchStatistic;
 import step3.lotto.domain.lotto.Winnings;
 
@@ -12,7 +11,6 @@ import step3.lotto.domain.lotto.Winnings;
 public class LottoService {
 
     public MatchStatistic play(Customer customer, Winnings winnings) {
-        Lottos lottos = customer.getLottos();
-        return lottos.match(winnings);
+        return winnings.match(customer.getLottos());
     }
 }
