+++ src/test/java/step3/lotto/domain/lotto/WinningsTest.java
@@ -2,6 +2,7 @@ package step3.lotto.domain.lotto;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
+import static org.junit.jupiter.api.Assertions.assertAll;
 import static step3.lotto.domain.lotto.Winnings.BONUS_NUMBER_ALREADY_EXIST_IN_WINNING_LOTTO_ERROR;
 
 import java.util.Arrays;
@@ -27,6 +28,37 @@ public class WinningsTest {
 
         // Then
         assertThat(winnings).isNotNull();
+    }
+
+    @Test
+    @DisplayName("N번의 로또 게임의 당첨 통계 반환")
+    public void returnMatchResults_GivenLottos() {
+        // Given
+        Lottos lottos = new Lottos(Arrays.asList(
+            Lotto.of(Arrays.asList(1, 2, 3, 4, 5, 6)),
+            Lotto.of(Arrays.asList(1, 2, 3, 4, 5, 7)),
+            Lotto.of(Arrays.asList(1, 2, 3, 4, 5, 8)),
+            Lotto.of(Arrays.asList(1, 2, 3, 4, 7, 8)),
+            Lotto.of(Arrays.asList(1, 2, 3, 7, 8, 9)),
+            Lotto.of(Arrays.asList(1, 2, 7, 8, 9, 10)),
+            Lotto.of(Arrays.asList(1, 7, 8, 9, 10, 11)),
+            Lotto.of(Arrays.asList(7, 8, 9, 10, 11, 12))
+        ));
+        Lotto winningsLotto = Lotto.of(Arrays.asList(1, 2, 3, 4, 5, 6));
+        LottoNumber lottoNumber = LottoNumber.of(7);
+        Winnings winnings = Winnings.of(winningsLotto, lottoNumber);
+
+        // When
+        MatchStatistic matchStatistic = winnings.match(lottos);
+
+        // Then
+        assertAll(
+            () -> assertThat(matchStatistic.getFirstPlaceCount()).as("6개 일치 수").isEqualTo(1),
+            () -> assertThat(matchStatistic.getSecondPlaceCount()).as("5개 일치 수").isEqualTo(1),
+            () -> assertThat(matchStatistic.getThirdPlaceCount()).as("4개 일치 수").isEqualTo(1),
+            () -> assertThat(matchStatistic.getForthPlaceCount()).as("3개 일치 수").isEqualTo(1),
+            () -> assertThat(matchStatistic.getRateOfProfit()).as("수익률").isEqualTo(250569.37)
+        );
     }
 
     @Test
